Bootstrap environment if running locally.

+++ scripts/run_tests.sh
@@ -5,6 +5,15 @@ DM_ENVIRONMENT=${1:-local}
 bundle install
 . "./scripts/envs/${DM_ENVIRONMENT}.sh"
 
+./scripts/test_services.sh "$DM_ENVIRONMENT" || exit 1
+
+if [ "$DM_ENVIRONMENT" = "local" ]; then
+  echo -e "\033[0;34mBootstrapping local environment\033[0m"
+  ./nginx/bootstrap.sh
+  ./scripts/create_users.sh "$DM_ENVIRONMENT"
+  exit
+fi
+
 mkdir -p reports
 rm -f screenshot*
